PowerStats: extend the MAX_STATS_ARRAY_LENGTH
Shrink the bits of parcel version to extend the MAX_STATS_ARRAY_LENGTH
to prevent the OOB of MAX_STATS_ARRAY_LENGTH when the number of cpu frequencies
exceeds the original MAX_STATS_ARRAY_LENGTH.

Google: 3480310

+++ services/core/jni/com_android_server_power_stats_CpuPowerStatsCollector.cpp
@@ -113,7 +113,7 @@ static jlong nativeReadCpuStats(JNIEnv *env, [[maybe_unused]] jobject zis, jobje
 static int flatten(JNIEnv *env, const std::vector<std::vector<uint64_t>> &times,
                    jlongArray outArray) {
     ScopedLongArrayRW scopedOutArray(env, outArray);
-    const uint8_t scalingStepCount = scopedOutArray.size();
+    const uint32_t scalingStepCount = scopedOutArray.size();
     uint64_t *out = reinterpret_cast<uint64_t *>(scopedOutArray.get());
     uint32_t scalingStep = 0;
     for (const auto &subVec : times) {
@@ -136,9 +136,9 @@ static int combineByBracket(JNIEnv *env, const std::vector<std::vector<uint64_t>
                             jlongArray outBrackets) {
     ScopedLongArrayRW scopedOutBrackets(env, outBrackets);
     uint64_t *brackets = reinterpret_cast<uint64_t *>(scopedOutBrackets.get());
-    const uint8_t statsSize = scopedOutBrackets.size();
+    const uint32_t statsSize = scopedOutBrackets.size();
     memset(brackets, 0, statsSize * sizeof(uint64_t));
-    const uint8_t scalingStepCount = scopedScalingStepToPowerBracketMap.size();
+    const uint32_t scalingStepCount = scopedScalingStepToPowerBracketMap.size();
 
     uint32_t scalingStep = 0;
     for (const auto &subVec : times) {
